services/api/pyproject.toml
[tool.poetry]
name = "backpack-api"
version = "0.1.0"
description = ""
authors = ["Chase <chase@theponti.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
flask-sqlalchemy = "^2.4.1"
psycopg2-binary = "^2.8.4"
gunicorn = "^20.0.4"
coverage = "^5.0.3"
pytest = "^5.3.5"
flask-migrate = "^2.5.2"
graphene = "^2.1.8"
graphene_sqlalchemy = "^2.2.2"
Flask-GraphQL = "^2.0.1"
Flask-Testing = "^0.7.1"
requests = "^2.23.0"
werkzeug = "0.16.0"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
